Added data extraction/wrangling scripts and data files
`import_crime.py` takes some Socrata credentials (`app_token` is not
strictly necessary, but it avoids issues with throttling), downloads
the raw data for 2010, and writes it to a file (`crimes_raw.csv`).

The IPython notebook contains the scripts for data manipulation and
cleaning. Crimes without a clearly-defined neighborhood are excluded
(as of this commit, the dataset shrinks from 21,089 to 20,794 rows).
Using data from the City of Cincinnati and the 2010 Census, crime and
population counts are found by neighbourhood, resulting in the final
data file: `crimes_clean_counts_and_pops.csv`. (`crimes_clean.csv` is
also saved for record-keeping and double-checking.)