Moved to higher-quality geographic data
The previous data had a lot of gaps, overlaps, and intersections which
were causing nightmares in the geospatial analysis. The new geographic
data also re-separates North Avondale and Paddock Hills. The updated
map may be somewhat anachronistic, but contiguity relations should
be the same, while areas and perimeters appeared to be similar from a
cursory investigation.