Re-scaled crime report counts
Did some exploratory data analysis and realised the first half of 2010
was pretty much empty. Ran a simple linear regression to get an estimate
for the total counts (model formula and results are included in
`analysis/report.Rmd`). Updated the source and output data files.

Also improved the data import script to take the year as an argument.