Added the simulation code and the generated output
The notebook includes some additional data preparation but is
primarily the code for actually running the simulation. This builds
off of the general approach used in DEAP, though most of the actual
operators and helper functions needed to be re-implemented from scratch
to account for geospatial constraints.

The final output from the simulation is saved in `output.csv`, which
contains the zone memberships for each unit and their related fitness
scores. Out of an abundance of caution, the block responsible for
actually saving the results to disk is commented out.