Added the first round of analysis
The notebook contains the first pass of analysis on the simulation
output, including Shapiro-Wilk and Anderson-Darling normality tests
for the compactness, population counts, and crime counts. The final
diagnostic used is the Adjusted Rand Index, which from what I've
reviewed is somewhat novel (that said, I did not conduct anything near
an exhaustive literature review).

`evaluate.py` contains the same fitness functions used in the
simulation notebook. I tried refactoring, but I'm not particularly
adept with imports in Python. (Unlike Node.js which allows for explicit
pathing through directories, Python assumes everything is being
imported upward?) I'll come back around to refactoring this so I'm
not risking any unnoticed changes from one location to the next.

Finally, there are an assortment of images to be used in the final
report, which I've included in the repo for the sake of posterity.