Wrote first draft of report and supporting files
`before_body.tex` is a workaround for getting a proper title page out
of the Rmarkdown document template, which is admittedly a bit of a hack
in and of itself.

`references.bib` are, well, my references.

Also updated the handling of map rendering in the map wrangling
notebook to reflect the project styles.